samba/requirements.txt
openai==1.54.5
python-dotenv==1.0.1
scikit-learn==1.5.2
